requirements.txt
django>=4.2,<5.0
pillow>=10.0
gspread>=5.12
google-auth>=2.23
gunicorn>=21.2
django-cors-headers>=4.3
python-dotenv>=1.0
whitenoise>=6.6
requests>=2.31
opencv-python-headless>=4.8.0
numpy>=1.24.0